requirements.txt
discord.py[voice]>=2.3.2
youtube-dl>=2021.12.17
PyNaCl>=1.5.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.16.1
aiohttp>=3.8.5
yt-dlp>=2023.12.30
